+++ AudioVideoCode/08-rgb-to-yuv-cpp/src/entity/RGBPixel.cpp
@@ -1,6 +1,6 @@
 #include "../../include/entity/RGBPixel.h"
 
-RGBPixel::RGBPixel(uint8_t r, uint8_t g, uint8_t b)
+RGBPixel::RGBPixel(uint32_t r, uint32_t g, uint32_t b)
 {
     this->red = r;
     this->green = g;
